requirements.txt
pyspark==3.5.0
apache-airflow-providers-apache-spark==4.7.0
